Improve calculation steps in renderer
This commit improves the implementation in 2 aspects: Firstly, it
reduces the amount of computations in the function trinity() by
eliminating the step of transforming the vertices of triangle to screen
space. Secondly, it updates the method of computing the coordinate of a
pixel in order to keep the screen ratio correct and not to squash the
shape of objects being rendered.