Add Verilator simulator
This commit imports an efficient Verilog-to-C++ translator called
Verilator that can generates all the required C++ files for runtime
simulation. A C++ program is also added and can be used with the help of
the generated files of Verilator to simulates the expected execution
results of Trinity in a SDL window.