Add the shader class
This commit adds a new class called Shader, which includes the
definition of several parameters of the VGA signal. With this class, all
those constants can be used freely in anywhere in this project.
Additionally, this commit also tweaked the fields of the classes such as
ScreenPos (the fact is that it's not a real class but an alias of UVec2
however) due to the introduction of the Screen class.